Work-arounds of Kelvin cmd line sim + full interrupt handlers
Coded and ran in simulation real interrupt handling, not just looking at
the interrupt pending/"flag" register (IFR).

In Kelvin cmd line sim, the interrupt mux model does not appear to
let the interrupts through if you just use the default setting of the
register even though the register reads correctly.  Work-around
is to read and write back MUXH.  (MUXL probably works too.)  Of course
if you change the MUX settings that works too.  This workaround is not
needed in CCS simulator.

Also the MUX model does not seem to model routing an input interrupt to
more than one device interrupt.  This is probably only interesting for
cases like this but it does mismatch the CCS simulator and (I presume) HW.