support userspace builds with GCC
This patch adds the ability to build userspace components with GCC. This is
controlled by the BUILD_USERSPACE_WITH_GCC variable. At this time, the uclibc
sources are different for GCC builds and TI toolchain builds. This patch does
not contain the uclibc sources for the GCC build. Those sources are in the
Code Sourcery toolchain source tarball which should be unpacked in the top
level directory (LINUX_C6X_TOP_DIR).

Signed-off-by: Mark Salter <msalter@redhat.com>