build up a GNU-toolchain based SDK
A previous patch added support for building userspace components with GCC. It
did this by using a pre-packaged GNU toolchain. This patch uses the pre-packaged
GNU-toolchain as a starting point to build up an SDK which contains new uClibc
libraries and headers built from the project sources. This ensures that the
toolchain (SDK) libraries match the runtime rootfs libraries. Also, it allows
for easy addition of extra libraries to the SDK.

Signed-off-by: Mark Salter <msalter@redhat.com>