Fixing uClibc build issue
When all of the repositories are cloned and started a product build from the c66x branch, uClibc
failed to build. It failed in the install phase where uClibc is expecting $(SYSROOT_DIR)/lib
and $(SYSROOT_DIR)/usr/lib directories to be created. So add rule to create them in Makefile.

Signed-off-by: Muralidharan Karicheri <m-karicheri2@ti.com>