elf loader updates to build from linux-c6x-project
Also following changes done.

The syslink scripts are modified to use the elf loader
in the filesystem. The elf loader is present in the
/usr/bin/dl6x.gcc for demo mcsdk rootfs. The
syslink scripts uses this as the default location.
If user is using any other rootfs, user needs to provide
the path of the dl6x.gcc as first argument of the
syslink scripts.

Signed-off-by: Muralidharan Karicheri <m-karicheri2@ti.com>