gcc-install.sh: major refactor to better handle multiple public releases
The previous commit to this file did not actually make the 4.5-123 version downloadable except inside TI.
Actaully doing so in the old structure would have been a major pain.
This is a major refactor to make it more maintainable with multiple public releases.
It also adds a few features.
It should be noted that 4.5-123 is not actually public at this point so it may need some adjusting when it is.

Signed-off-by: Bill Mills <wmills@ti.com>