./prj config: install to $TOP/opt; add install of BIOS, XDC, XDAIS
If we install anything, install it to project's $TOP/opt dir.
This includes gcc and uclibc
If uclibc came from a source tar file it will not be in $TOP/opt but in $TOP
Thus it is always safe to rm -rf $TOP/opt (provided TOP is set correctly ;) )

Now knows how to install BIOS, XDC, and XDAIS
This uses expanded generic component handling functions
Converted IPC to use these generic component functions as well.

Signed-off-by: Bill Mills <wmills@ti.com>