Bootblobs: Move Bootblob templates to shell and allow use in product bundles
Previously the predefined bootblobs were a Makefile facility and only available to source users
With this commit:
- these predefined bootblobs are called bootblob templates
- are implemented in shell fragments
- are copied to the product directory
- are invoked by a new mode of the bootblob script

This allows the creation of bootblobs to be defered until the end user has downloaded a
binary release.  This means she does not need to download a bunch of redundent bits and
(eventually) the created products can be customized by user specific settings.

It is still possible to specify templates that should be run at make product time in setenv.

This commit also quiets the default bootblob operations a bit by default.

Signed-off-by: Bill Mills <wmills@ti.com>