bootblob templates: Fix IPADDR for evmc6474-lite
The templates look for an env variable of the form <evm name>_IPADDR.
However the dash "-" in evmc6474-lite is not a valid character for an env variable name.
This caused invalid setting for ip= kernel command line param.
So convert all dashes to underscores in the evm name before using in this context.
(The script has always uppercased the evm name also)

So to set just the evmc6474-lite to use dhcp you would:
export EVMC6474_LITE_IPADDR=dhcp
[or any other valid setting for the ip= kernel command line param]

If an evm specific value is not found, the value of IPADDR is used.
If that is not found then the default is "dhcp"

Fixes bug #23

Signed-off-by: Bill Mills <wmills@ti.com>